Adds Android XML Composition
(Works with Layout Files, and should work with most others)

Includes the ability to use hooking comments to place widgets in a
branch before or after other widgets in the same branch. If a widget id
is not added, it will either prepend or append to that branch.

Example of placing a button before another widget (android id =
@+id/l_wordpressdotcom), just change to:

<!-- @start before android:id="@+id/l_wordpressdotcom"-->
    <Button
      android:layout_width="wrap_content"
      android:layout_height="wrap_content"
      android:id="@+id/wordpressdotcom"
      android:text="@string/get_free_blog"
      android:layout_centerHorizontal="true"
      android:layout_below="@id/l_wordpressdotcom"
      android:background="@drawable/wp_button" />
<!-- @end before android:id="@+id/l_wordpressdotcom"-->