Change external line based merge tool to git merge-file.
Git is way more common than RCS by now, so it is more likely to be
installed on modern operating systems. With 'merge-file', it provides a
minimal clone of RCS merge. Therefore, we can get rid of the
additional dependency to the rather ancient RCS.

I also looked into using 'diff3 --merge -E', but found that it uses an
inferior merge algorithm that produces larger conflicts.