Fix for --filemerge.
According to the usage information, the argument --filemerge takes a
triple of variants (mine base other) and is supposed to be an
alternative to providing an expression file with -e.

However, this didn't really work so far, as the three arguments were
never used. FSTMerge just ignored them and tried to use a non-existing
expression file, resulting in a crash.

This is a workaround that generates a temporary expression file
containing the three arguments from --filemerge.