Fix overridden javac in build file.
ant doesn't like overriding the javac task any more and prints a warning
message during each build. It's (for now) harmless but annoying. I went
with the easiest fix, which is simply renaming the task (to my.javac in
our case). A test build confirmed that the warning message no longer
appears.